docker/server/requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2==2.8.6
